+++ src/server/HistoryHandler.java
@@ -1,10 +1,8 @@
-package Server;
+package server;
 
 import com.sun.net.httpserver.HttpExchange;
 import service.TaskManager;
-
 import java.io.IOException;
-
 
 
 public class HistoryHandler extends BaseHttpHandler {
