+++ src/server/HttpTaskServer.java
@@ -1,4 +1,4 @@
-package Server;
+package server;
 
 
 import adapters.LocalDateTimeAdapter;
