+++ src/test/server/HttpTaskManagerTasksTest.java
@@ -1,7 +1,7 @@
 package test.server;
 
 
-import Server.HttpTaskServer;
+import server.HttpTaskServer;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import model.Status;
